Add files via upload
This is the test file using watershed to use the most differentiable sections of my incoming picture to simplify it. I used this to write code to make weights of certain case options heavier or lighter. In other words, I chose the brightest and darkest parts of the picture and made those far brighter and far darker using the watershed function to simplify the picture, then ran code to analyze the simplified picture. In this file is only the test function that I used to learn how to use the watershed function.